requirements.txt
Flask==3.1.0
boto3==1.35.0
aws-cdk-lib==2.160.0
constructs>=10.0.0
pydantic==2.9.0
python-dotenv==1.0.1
websocket-client==1.8.0
requests==2.32.3
flask-socketio==5.3.6
eventlet==0.33.3
paho-mqtt==1.6.1
aiortc==1.6.0
websockets==11.0.3
pytest==8.3.0
